Cleanup/Modify Build System (Credit CM)

+++ envsetup.sh
@@ -437,10 +437,6 @@ function add_lunch_combo()
 }
 
 # add the default one here
-#add_lunch_combo full-eng
-#add_lunch_combo full_x86-eng
-#add_lunch_combo vbox_x86-eng
-#add_lunch_combo full_mips-eng
 
 function print_lunch_menu()
 {
@@ -448,7 +444,11 @@ function print_lunch_menu()
     echo
     echo "You're building on" $uname
     echo
-    echo "Breakfast menu... pick a combo:"
+    if [ "z${LIQUID_DEVICES_ONLY}" != "z" ]; then
+       echo "Breakfast menu... pick a combo:"
+    else
+       echo "Lunch menu... pick a combo:"
+    fi
 
     local i=1
     local choice
@@ -469,7 +469,7 @@ function brunch()
 {
     breakfast $*
     if [ $? -eq 0 ]; then
-        export FAST_BUILD=1
+        export LIQUID_FAST_BUILD=1
         mka liquid
     else
         echo "No such item in brunch menu. Try 'breakfast'"
@@ -500,7 +500,7 @@ function breakfast()
             # A buildtype was specified, assume a full device name
             lunch $target
         else
-            # This is probably just the CM model name
+            # This is probably just the LIQUID model name
             lunch liquid_$target-userdebug
         fi
     fi
@@ -1270,6 +1270,7 @@ function installboot()
         return 1
     fi
     PARTITION=`grep "^\/boot" $OUT/recovery/root/etc/recovery.fstab | awk {'print $3'}`
+    PARTITION_TYPE=`grep "^\/boot" $OUT/recovery/root/etc/recovery.fstab | awk {'print $2'}`
     if [ -z "$PARTITION" ];
     then
         echo "Unable to determine boot partition."
@@ -1288,7 +1289,12 @@ function installboot()
         do
             adb push $i /system/lib/modules/
         done
-        adb shell dd if=/cache/boot.img of=$PARTITION
+        if [ "$PARTITION_TYPE" == "mtd" ];
+        then
+            adb shell flash_image $PARTITION /cache/boot.img
+        else
+            adb shell dd if=/cache/boot.img of=$PARTITION
+        fi
         adb shell chmod 644 /system/lib/modules/*
         echo "Installation complete."
     else
